@@ -5,7 +5,7 @@ publish_to: none
 version: 0.0.2
 
 environment:
-  sdk: '>=3.5.4 <4.0.0'
+  sdk: '>=3.6.0 <4.0.0'
 
 dependencies:
   http: ^1.2.2
