@@ -2,18 +2,18 @@ name: tentura
 description: 'Social network for communities'
 publish_to: 'none'
 
-version: 0.9.5
+version: 0.9.6
 
 environment:
-  sdk: '>=3.5.4 <4.0.0'
-  flutter: ^3.24.5
+  sdk: '>=3.6.0 <4.0.0'
+  flutter: ^3.27.0
 
 dependencies:
   flutter:
     sdk: flutter
   nil: ^1.1.1
   uuid: ^4.4.2
-  intl: ^0.19.0
+  intl: ^0.20.1
   get_it: ^8.0.2
   logger: ^2.5.0
   latlong2: ^0.9.1
